requirements.txt
pygame==2.5.0
python-chess==1.999 